CDI integration. Vault, ObjectStorage, Metrics
