+++ bom/pom.xml
@@ -916,16 +916,6 @@
                 <artifactId>helidon-integrations-oci-cdi</artifactId>
                 <version>${helidon.version}</version>
             </dependency>
-            <dependency>
-                <groupId>io.helidon.integrations.oci</groupId>
-                <artifactId>helidon-integrations-oci-vault-blocking</artifactId>
-                <version>${helidon.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>io.helidon.integrations.oci</groupId>
-                <artifactId>helidon-integrations-oci-objectstorage-blocking</artifactId>
-                <version>${helidon.version}</version>
-            </dependency>
             <!-- OpenAPI support -->
             <dependency>
                 <groupId>io.helidon.openapi</groupId>
