+++ examples/integrations/oci/objectstorage-cdi/pom.xml
@@ -48,7 +48,7 @@
         </dependency>
         <dependency>
             <groupId>io.helidon.integrations.oci</groupId>
-            <artifactId>helidon-integrations-oci-objectstorage-blocking</artifactId>
+            <artifactId>helidon-integrations-oci-objectstorage</artifactId>
         </dependency>
         <dependency>
             <groupId>io.helidon.config</groupId>
