+++ examples/integrations/oci/vault-cdi/src/main/resources/application.yaml
@@ -5,11 +5,16 @@
 server:
   port: 8080
 
-oci:
+app:
   vault:
     # Vault OCID (the vault you want to use for this example
     vault-ocid: "${oci.properties.vault-ocid}"
     compartment-ocid: "${oci.properties.compartment-ocid}"
     encryption-key-ocid: "${oci.properties.vault-key-ocid}"
     signature-key-ocid: "${oci.properties.vault-rsa-key-ocid}"
-    cryptographic-endpoint: "${oci.properties.cryptographic-endpoint}"
+
+oci:
+  custom:
+    vault:
+      # named OCI configuration
+      cryptographic-endpoint: "${oci.properties.cryptographic-endpoint}"
