+++ integrations/oci/cdi/src/main/java/io/helidon/integrations/oci/cdi/OciCdiExtension.java
@@ -16,15 +16,23 @@
 
 package io.helidon.integrations.oci.cdi;
 
+import java.lang.annotation.Annotation;
+import java.lang.reflect.Type;
+import java.util.HashSet;
 import java.util.List;
 import java.util.ServiceLoader;
+import java.util.Set;
 
-import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.event.Observes;
+import javax.enterprise.inject.literal.NamedLiteral;
 import javax.enterprise.inject.spi.AfterBeanDiscovery;
+import javax.enterprise.inject.spi.Annotated;
 import javax.enterprise.inject.spi.CDI;
 import javax.enterprise.inject.spi.Extension;
-import javax.inject.Singleton;
+import javax.enterprise.inject.spi.InjectionPoint;
+import javax.enterprise.inject.spi.ProcessBean;
+import javax.enterprise.inject.spi.ProcessInjectionPoint;
+import javax.inject.Named;
 
 import io.helidon.common.serviceloader.HelidonServiceLoader;
 import io.helidon.config.Config;
@@ -33,36 +41,152 @@ import io.helidon.integrations.oci.connect.spi.InjectionProvider;
 import io.helidon.microprofile.cdi.RuntimeStart;
 
 public class OciCdiExtension implements Extension {
+    private final Set<Type> supportedTypes = new HashSet<>();
+    private final Set<String> requiredNames = new HashSet<>();
+    private final List<InjectionProvider> providers;
+
     private Config config;
+
+    public OciCdiExtension() {
+        providers = HelidonServiceLoader
+                .builder(ServiceLoader.load(InjectionProvider.class))
+                .build()
+                .asList();
+
+        for (InjectionProvider provider : providers) {
+            provider.injectables()
+                    .stream()
+                    .map(InjectionProvider.InjectionType::injectedType)
+                    .forEach(supportedTypes::add);
+        }
+    }
 
     private void configure(@Observes @RuntimeStart Config config) {
         this.config = config.get("oci");
     }
 
-    void registerProducers(@Observes AfterBeanDiscovery abd) {
-        abd.addBean()
-                .types(OciRestApi.class)
-                .beanClass(OciRestApi.class)
-                .scope(Singleton.class)
-                .createWith(ignored -> OciRestApi.create(config));
+    /**
+     * Add internal qualifier.
+     *
+     * @param event CDI event
+     */
+    void updateInjectionPoints(@Observes ProcessInjectionPoint<?, ?> event) {
+        InjectionPoint injectionPoint = event.getInjectionPoint();
+        Annotated annotated = injectionPoint.getAnnotated();
 
-        List<InjectionProvider> providers = HelidonServiceLoader
-                .builder(ServiceLoader.load(InjectionProvider.class))
-                .build()
-                .asList();
+        Type type = injectionPoint.getType();
+        if (supportedTypes.contains(type)) {
+            Named name = annotated.getAnnotation(Named.class);
 
-        for (InjectionProvider<?> provider : providers) {
-            abd.addBean()
-                    .types(provider.types())
-                    .beanClass(OciCdiExtension.class)
-                    .scope(ApplicationScoped.class)
-                    .createWith(ignored -> {
-                        OciRestApi restApi = CDI.current()
-                                .select(OciRestApi.class)
-                                .get();
+            OciInternal internal = OciInternal.Literal
+                    .create((name == null ? "" : name.value()));
 
-                        return provider.createInstance(restApi, config);
-                    });
+            event.configureInjectionPoint()
+                    .addQualifier(internal);
         }
+    }
+
+    /**
+     * Collect injection points that are valid.
+     *
+     * @param event CDI event
+     */
+    void processInjectionPointsFromEnabledBeans(@Observes ProcessBean<?> event) {
+        for (InjectionPoint injectionPoint : event.getBean().getInjectionPoints()) {
+            Set<Annotation> qualifiers = injectionPoint.getQualifiers();
+            for (Annotation qualifier : qualifiers) {
+                if (qualifier.annotationType().equals(OciInternal.class)) {
+                    OciInternal value = (OciInternal) qualifier;
+
+                    requiredNames.add(value.value());
+                    break;
+                }
+            }
+        }
+    }
+
+    void registerProducers(@Observes AfterBeanDiscovery abd) {
+        if (requiredNames.contains("")) {
+            requiredNames.add("default");
+        } else if (requiredNames.contains("default")) {
+            requiredNames.add("");
+        }
+        boolean addDefault = true;
+
+        for (String name : requiredNames) {
+            Named named = NamedLiteral.of(name);
+
+            if (name.isEmpty() || name.equals("default")) {
+                if (addDefault) {
+                    // default
+                    abd.addBean(new QualifiedBean<>(OciCdiExtension.class,
+                                                    OciRestApi.class,
+                                                    () -> {
+                                                        if (config.get("default").exists()) {
+                                                            return OciRestApi.create(config.get("default"));
+                                                        } else {
+                                                            return OciRestApi.create(config);
+                                                        }
+                                                    }));
+
+                    abd.addBean(new QualifiedBean<>(OciCdiExtension.class,
+                                                    OciRestApi.class,
+                                                    Set.of(NamedLiteral.of(""), NamedLiteral.of("default")),
+                                                    () -> {
+                                                        if (config.get("default").exists()) {
+                                                            return OciRestApi.create(config.get("default"));
+                                                        } else {
+                                                            return OciRestApi.create(config);
+                                                        }
+                                                    }));
+                    addDefault = false;
+                }
+            } else {
+                // named
+                abd.addBean(new QualifiedBean<>(OciRestApi.class,
+                                                OciRestApi.class,
+                                                Set.of(named),
+                                                () -> {
+                                                    System.out.println("Creating OCI REST API from config: " + config.get(name)
+                                                            .key());
+                                                    return OciRestApi.create(config.get(name));
+                                                }
+                ));
+            }
+        }
+
+        // now we need to add the full combination of names and types
+        for (String name : requiredNames) {
+            OciInternal.Literal qualifier = OciInternal.Literal.create(name);
+            Named named = NamedLiteral.of(name);
+
+            // for each name
+            for (InjectionProvider provider : providers) {
+                // for each provider
+                for (InjectionProvider.InjectionType<?> injectable : provider.injectables()) {
+                    // for each type
+                    abd.addBean(new QualifiedBean<>(OciCdiExtension.class,
+                                                    (Class<Object>) injectable.injectedType(),
+                                                    Set.of(named, qualifier),
+                                                    () -> {
+                                                        OciRestApi restApi = CDI.current()
+                                                                .select(OciRestApi.class, named)
+                                                                .get();
+
+                                                        return injectable.createInstance(restApi, producerConfig(name));
+                                                    }));
+                }
+            }
+        }
+    }
+
+    private Config producerConfig(String name) {
+        if (name.isEmpty() || name.equals("default")) {
+            if (config.get("default").exists()) {
+                return config.get("default");
+            }
+            return config;
+        }
+        return config.get(name);
     }
 }
