+++ integrations/oci/telemetry/src/main/java/io/helidon/integrations/oci/telemetry/OciMetrics.java
@@ -16,56 +16,12 @@
 
 package io.helidon.integrations.oci.telemetry;
 
-import java.util.function.Consumer;
-
-import io.helidon.common.reactive.Single;
-import io.helidon.config.Config;
-import io.helidon.integrations.oci.connect.OciRestApi;
-
 /**
- * Reactive APIs for OCI Metrics.
+ * Blocking APIs for OCI Metrics.
  */
 public interface OciMetrics {
-    /**
-     * Version of API supported by this client.
-     */
-    String API_VERSION = "20180401";
-    /**
-     * Host name prefix.
-     */
-    String API_HOST_PREFIX = "telemetry-ingestion";
-    /**
-     * Host format of API server.
-     */
-    String API_HOST_FORMAT = "%s://%s.%s.%s";
-
-    /**
-     * Create a new fluent API builder for OCI metrics.
-     *
-     * @return a new builder
-     */
-    static Builder builder() {
-        return new Builder();
-    }
-
-    /**
-     * Create OCI metrics using the default {@link io.helidon.integrations.oci.connect.OciRestApi}.
-     *
-     * @return OCI metrics instance connecting based on {@code DEFAULT} profile
-     */
-    static OciMetrics create() {
-        return builder().build();
-    }
-
-    /**
-     * Create OCI metrics based on configuration.
-     *
-     * @param config configuration on the node of OCI configuration
-     * @return OCI metrics instance configured from the configuration
-     * @see io.helidon.integrations.oci.telemetry.OciMetrics.Builder#config(io.helidon.config.Config)
-     */
-    static OciMetrics create(Config config) {
-        return builder().config(config).build();
+    static OciMetrics create(OciMetricsRx reactive) {
+        return new OciMetricsImpl(reactive);
     }
 
     /**
@@ -85,80 +41,7 @@ public interface OciMetrics {
      * <a href="https://docs.oracle.com/iaas/Content/Monitoring/Concepts/monitoringoverview.htm#concepts">Monitoring Concepts</a>.
      *
      * @param request metric request
-     * @return future metric response
+     * @return metric response
      */
-    Single<PostMetricData.Response> postMetricData(PostMetricData.Request request);
-
-    class Builder implements io.helidon.common.Builder<OciMetrics> {
-        private final OciRestApi.Builder accessBuilder = OciRestApi.builder();
-
-        private String apiVersion = API_VERSION;
-        private String hostPrefix = API_HOST_PREFIX;
-        private String endpoint;
-
-        private Builder() {
-        }
-
-        @Override
-        public OciMetrics build() {
-            return new OciMetricsImpl(this);
-        }
-
-        /**
-         * Update from configuration. The configuration must be located on the {@code OCI} root configuration
-         * node.
-         *
-         * @param config configuration
-         * @return updated metrics builder
-         */
-        public Builder config(Config config) {
-            accessBuilder.config(config);
-            config.get("metrics.host-prefix").asString().ifPresent(this::hostPrefix);
-            config.get("metrics.endpoint").asString().ifPresent(this::endpoint);
-            config.get("metrics.api-version").asString().ifPresent(this::apiVersion);
-            return this;
-        }
-
-        public Builder hostPrefix(String prefix) {
-            this.hostPrefix = prefix;
-            return this;
-        }
-
-        public Builder endpoint(String endpoint) {
-            this.endpoint = endpoint;
-            return this;
-        }
-
-        public  Builder apiVersion(String apiVersion) {
-            this.apiVersion = apiVersion;
-            return this;
-        }
-
-        /**
-         * Update the rest access builder to modify defaults.
-         *
-         * @param builderConsumer consumer of the builder
-         * @return updated metrics builder
-         */
-        public Builder updateRestAccess(Consumer<OciRestApi.Builder> builderConsumer) {
-            builderConsumer.accept(accessBuilder);
-            return this;
-        }
-
-        String apiVersion() {
-            return apiVersion;
-        }
-
-        OciRestApi restAccess() {
-            return accessBuilder.build();
-        }
-
-        String hostPrefix() {
-            return hostPrefix;
-        }
-
-        String endpoint() {
-            return endpoint;
-        }
-    }
+    PostMetricData.Response postMetricData(PostMetricData.Request request);
 }
