+++ integrations/oci/telemetry/src/main/resources/META-INF/services/io.helidon.integrations.oci.connect.spi.InjectionProvider
@@ -1,0 +1,1 @@
+io.helidon.integrations.oci.telemetry.OciTelemetryInjectionProvider
