--- integrations/oci/vault-blocking/src/main/java/module-info.java
@@ -1,13 +1,0 @@
-import io.helidon.integrations.oci.vault.blocking.OciBlockingVaultInjectionProvider;
-
-module io.helidon.integrations.oci.vault.blocking {
-    requires io.helidon.integrations.oci.connect;
-    requires io.helidon.integrations.oci.vault;
-    requires io.helidon.config;
-    requires io.helidon.integrations.common.rest;
-
-    exports io.helidon.integrations.oci.vault.blocking;
-
-    provides io.helidon.integrations.oci.connect.spi.InjectionProvider
-        with OciBlockingVaultInjectionProvider;
-}
