--- integrations/oci/vault-blocking/src/main/resources/META-INF/services/io.helidon.integrations.oci.connect.spi.InjectionProvider
@@ -1,1 +1,0 @@
-io.helidon.integrations.oci.vault.blocking.OciBlockingVaultInjectionProvider
