+++ integrations/oci/vault/src/main/java/io/helidon/integrations/oci/vault/OciVaultInjectionProvider.java
@@ -1,23 +1,33 @@
 package io.helidon.integrations.oci.vault;
 
-import java.lang.reflect.Type;
-import java.util.Set;
+import java.util.LinkedList;
+import java.util.List;
 
-import io.helidon.config.Config;
-import io.helidon.integrations.oci.connect.OciRestApi;
 import io.helidon.integrations.oci.connect.spi.InjectionProvider;
 
-public class OciVaultInjectionProvider implements InjectionProvider<OciVault> {
-    @Override
-    public Set<Type> types() {
-        return Set.of(OciVault.class);
+public class OciVaultInjectionProvider implements InjectionProvider {
+    private static final List<InjectionType<?>> INJECTABLES;
+
+    static {
+        List<InjectionType<?>> injectables = new LinkedList<>();
+
+        injectables.add(InjectionType.create(OciVaultRx.class,
+                                             (restApi, config) -> OciVaultRx.builder()
+                                                     .restApi(restApi)
+                                                     .config(config)
+                                                     .build()));
+
+        injectables.add(InjectionType.create(OciVault.class,
+                                             (restApi, config) -> OciVault.create(OciVaultRx.builder()
+                                                                                          .restApi(restApi)
+                                                                                          .config(config)
+                                                                                          .build())));
+
+        INJECTABLES = List.copyOf(injectables);
     }
 
     @Override
-    public OciVault createInstance(OciRestApi restApi, Config ociConfig) {
-        return OciVault.builder()
-                .restApi(restApi)
-                .config(ociConfig)
-                .build();
+    public List<InjectionType<?>> injectables() {
+        return INJECTABLES;
     }
 }
