+++ integrations/oci/vault/src/main/java/io/helidon/integrations/oci/vault/OciVaultSecurityService.java
@@ -33,6 +33,6 @@ public class OciVaultSecurityService implements SecurityProviderService {
 
     @Override
     public SecurityProvider providerInstance(Config config) {
-        return new OciVaultSecurityProvider(OciVault.create(config));
+        return new OciVaultSecurityProvider(OciVaultRx.create(config));
     }
 }
