+++ integrations/oci/vault/src/main/resources/META-INF/services/io.helidon.integrations.oci.connect.spi.InjectionProvider
